feat!: Disable coverage maps unless specified with an option
Computing diff offsets is a pretty heavy process, especially for large files.
Therefore from now on we'll only do this if the importer has explicitly stated
it wants coverage maps to be generated.

BREAKING CHANGE: importer.getCoverageMap now requires
`generateCoverageMap: true` to be passed as an option to the importer.
See README.md for an example.