feat: Better error messages when a network error occurs
Previously fetching file:// imports that do not exist on disk, or
when fetching urls that respond with a non-ok status code, Deno
would throw an error. But from this error it was not immediately clear
it was caused by the importer, and no url was printed displaying the
cause of the failed import. This change handles both cases.

Fixes #3